python-dashboard/requirements.txt
tk
flask
matplotlib
pygame
PyQt5
paho-mqtt
Pillow